project/requirements.txt
jqdatasdk==1.8.47
flask>=2.0.0
pyyaml>=5.4.0
requests>=2.25.0
pytest>=7.0.0
